Extract diagnostics into core and features portion
Parsing, storing and looking up diagnostics is core
Displaying diagnostics handled in features
Also introduces panels
