+++ core/panels.py
@@ -1,0 +1,24 @@
+import sublime
+
+OUTPUT_PANEL_SETTINGS = {
+    "auto_indent": False,
+    "draw_indent_guides": False,
+    "draw_white_space": "None",
+    "gutter": False,
+    'is_widget': True,
+    "line_numbers": False,
+    "margin": 3,
+    "match_brackets": False,
+    "scroll_past_end": False,
+    "tab_size": 4,
+    "translate_tabs_to_spaces": False,
+    "word_wrap": False
+}
+
+
+def create_output_panel(window: sublime.Window, name: str) -> sublime.View:
+    panel = window.create_output_panel(name)
+    settings = panel.settings()
+    for key, value in OUTPUT_PANEL_SETTINGS.items():
+        settings.set(key, value)
+    return panel
